DB 커넥션 만들기의 독립
