+++ jin/README.md
@@ -56,3 +56,18 @@
 ##### DB 커넥션 만들기의 독립
 - UserDao 소스코드를 제공하지 않고도 고객이 원하는 DB 커넥션 생성 방식을 적용할 수 있을까?
 - 상속을 통한 확장
+  - DB 커넥션 연결이라는 관심을 상속을 통해 서브클래스로 분리
+  - UserDao: add(), get(), getConnection()
+  - NUserDao: getConnection()
+  - DUserDao: getConnection()
+- 템플릿 메소드 패턴(Template Method Pattern): 슈퍼클래스의 기능을 확장할 때 사용하는 가장 대표적인 방법
+  - 슈퍼클래스에 기본 로직흐름을 만들고, 서브클래스에서 각 메소드를 필요에 맞게 구현하는 방식
+  - 전체적으로 동일하면서 부분적으로 다른 구문으로 구성된 메소드의 코드 중복을 최소화하기에 유용
+- 팩토리 메소드 패턴(Factory Method Pattern)
+  - 서브클래스에서 구체적인 오브젝트 생성 방법을 결정하게 하는 것
+
+
+  
+## 1.3 DAO의 확장
+- 지금까지 데이터 엑세스 로직을 어떻게 만들 것인지, DB 연결을 어떤 방법으로 할 것인지 두 개의 관심을 기준으로 분리
+- 두 관심은 변화의 성격이 다름
