+++ jin/src/main/java/springbook/user/dao/DUserDao.java
@@ -1,0 +1,33 @@
+package springbook.user.dao;
+
+import springbook.user.domain.User;
+
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
+
+public class DUserDao extends UserDao {
+    public Connection getConnection() throws ClassNotFoundException, SQLException {
+        // D사 DB Connection 생성 코드
+        Class.forName("com.mysql.jdbc.Driver");
+        Connection c = DriverManager.getConnection("jdbc:mysql://localhost/springbook", "spring", "book");
+        return c;
+    }
+
+    public static void main(String[] args) throws SQLException, ClassNotFoundException {
+        User user = new User();
+        user.setId("naekang");
+        user.setName("김진호");
+        user.setPassword("haul1!");
+
+        UserDao userDao = new DUserDao();
+        userDao.add(user);
+
+        System.out.println(user.getId() + "등록 성공");
+
+        User user2 = userDao.get(user.getId());
+        System.out.println(user2.getName());
+        System.out.println(user2.getPassword());
+        System.out.println(user2.getId() + "조회 성공");
+    }
+}
