+++ jin/src/main/java/springbook/user/dao/UserDaoTest.java
@@ -1,25 +1,25 @@
-package springbook.user.dao;
-
-import springbook.user.domain.User;
-
-import java.sql.SQLException;
-
-public class UserDaoTest {
-    public static void main(String[] args) throws SQLException, ClassNotFoundException {
-        UserDao dao = new UserDao();
-
-        User user = new User();
-        user.setId("naekang");
-        user.setName("김진호");
-        user.setPassword("haul1!");
-
-        dao.add(user);
-
-        System.out.println(user.getId() + "등록 성공");
-
-        User user2 = new User();
-        System.out.println(user2.getName());
-        System.out.println(user2.getPassword());
-        System.out.println(user2.getId() + "조회 성공");
-    }
-}
+//package springbook.user.dao;
+//
+//import springbook.user.domain.User;
+//
+//import java.sql.SQLException;
+//
+//public class UserDaoTest {
+//    public static void main(String[] args) throws SQLException, ClassNotFoundException {
+//        UserDao userDao = new UserDao();
+//
+//        User user = new User();
+//        user.setId("naekang");
+//        user.setName("김진호");
+//        user.setPassword("haul1!");
+//
+//        dao.add(user);
+//
+//        System.out.println(user.getId() + "등록 성공");
+//
+//        User user2 = new User();
+//        System.out.println(user2.getName());
+//        System.out.println(user2.getPassword());
+//        System.out.println(user2.getId() + "조회 성공");
+//    }
+//}
